Fix AiToolInvocation context/hook export
