+++ packages/liveblocks-react-ui/src/primitives/AiMessage/contexts.ts
@@ -1,0 +1,17 @@
+import type { AiToolInvocationProps } from "@liveblocks/core";
+import { createContext, useContext } from "react";
+
+export const AiToolInvocationContext =
+  createContext<AiToolInvocationProps | null>(null);
+
+export function useAiToolInvocationContext() {
+  const context = useContext(AiToolInvocationContext);
+
+  if (context === null) {
+    throw new Error(
+      "This component must be used within a tool's render method."
+    );
+  }
+
+  return context;
+}
